WIP #32 - Ensure that web driver proxies expose the requested version
This means that factories which create proxies, all the way down to the
proxies themselves must include sufficient information to get the
originally requested flags.

Additionally, it has become apparent that the flags providers, types which
implement IGetsBrowserFlags, should use the browser identification (which
is rseolved with the best version name we have) instead of relying
directly on the web driver.